Criacao de teste e ajuste no codigo

+++ src/main/java/com/petshop/api/dto/ProductDTO.java
@@ -3,10 +3,14 @@ package com.petshop.api.dto;
 import com.petshop.domain.entity.Category;
 import jakarta.validation.constraints.NotNull;
 
+import java.math.BigDecimal;
+
 public record ProductDTO(
         Long id,
         @NotNull String name,
         Category category,
-        @NotNull double value,
+        @NotNull BigDecimal value,
         @NotNull int quantity
 ) {}
+
+// trocar o value para bigdecimal
